+++ docs/source/changes.md
@@ -121,6 +121,9 @@ authors:
   ({{ gh(482) }} by {{ authors.hoechenberger }})
 - Speed up report generation.
   ({{ gh(487) }} by {{ authors.hoechenberger }})
+- The new [`epochs_metadata_query][config.epochs_metadata_query] setting
+  allows to select epochs based on metadata query strings.
+  ({{ gh(495) }} by {{ authors.hoechenberger }} and {{ authors.agramfort }})
 
 ### Behavior changes
 
@@ -178,6 +181,9 @@ authors:
   size of the generated report, while maintaining a sufficiently detailed
   visualization.
   ({{ gh(488) }} by {{ authors.hoechenberger }})
+- In [`interactive`][config.interactive] mode, the Matplotlib `Agg` backend
+  will be used.
+  ({{ gh(497) }} by {{ authors.hoechenberger }})
 
 ### Code health